+++ src/MD5_fmt.c
@@ -124,7 +124,7 @@ static int get_hash_0(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xF;
 #else
 	init_t();
@@ -137,7 +137,7 @@ static int get_hash_1(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xFF;
 #else
 	init_t();
@@ -150,7 +150,7 @@ static int get_hash_2(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xFFF;
 #else
 	init_t();
@@ -163,7 +163,7 @@ static int get_hash_3(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xFFFF;
 #else
 	init_t();
@@ -176,7 +176,7 @@ static int get_hash_4(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xFFFFF;
 #else
 	init_t();
@@ -189,7 +189,7 @@ static int get_hash_5(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0xFFFFFF;
 #else
 	init_t();
@@ -202,7 +202,7 @@ static int get_hash_6(int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 	return ((MD5_word *)sout)[x+y*SIMD_COEF_32*4] & 0x7FFFFFF;
 #else
 	init_t();
@@ -298,7 +298,7 @@ static int cmp_one(void *binary, int index)
 #ifdef MD5_SSE_PARA
 	unsigned int x,y;
 	x = index&(SIMD_COEF_32-1);
-	y = index/SIMD_COEF_32;
+	y = (unsigned int)index/SIMD_COEF_32;
 
 	if(((unsigned int*)binary)[0] != ((unsigned int*)sout)[x+y*SIMD_COEF_32*4+0*SIMD_COEF_32])
 		return 0;
